requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
tenacity>=8.2.3

# AI & ML
torch>=2.1.0
transformers>=4.35.0
ollama>=0.1.0
whisperx>=3.1.0
elevenlabs>=0.2.24
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.2
nltk>=3.8.1
spacy>=3.7.2

# Database & Vector Store
chromadb>=0.4.15
redis>=5.0.1
asyncpg>=0.29.0
neo4j>=5.14.0
hnswlib>=0.7.0
sqlalchemy>=2.0.23
alembic>=1.12.1

# Security & Compliance
cryptography>=41.0.3
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
aiofiles>=23.2.1
PyJWT>=2.8.0

# Monitoring & Performance
prometheus-client>=0.17.1
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-prometheus>=0.54b1
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-redis>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
sentry-sdk>=1.35.0
structlog>=23.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0
pytest-mock>=3.12.0
pytest-env>=1.1.1
pytest-xdist>=3.5.0
coverage>=7.3.2

# Utilities
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.1
tenacity>=8.2.3
python-dateutil>=2.8.2
pytz>=2023.3
email-validator>=2.1.0
python-slugify>=8.0.1 